+++ lib/logd.dart
@@ -4,4 +4,4 @@ export 'src/handler/handler.dart';
 export 'src/logger/logger.dart' hide LogBuffer, LogEntry;
 export 'src/stack_trace/stack_trace.dart' hide CallbackInfo;
 export 'src/time/time.dart'
-    hide LocalTime, Time, TimezoneOffset, commonDSTRules, commonTimeZones;
+    hide LocalTime, Time, TimezoneOffset, commonDSTRules, commonTimezones;